refactor(reviewer): extract session driver
