+++ scripts/ci/module-test-impact.test.ts
@@ -62,6 +62,40 @@ describe('module test impact commands', () => {
     expect(plan.reasonChains).toContain('artifact_storage -> artifact_provenance')
   })
 
+  it.each(['src/main/reviewer/reviewer-session-driver.ts', 'src/shared/reviewer.ts'])(
+    'expands Reviewer changes through downstream consumers for %s',
+    (path) => {
+      const plan = createAffectedTestPlan([{ path, status: 'modified' }], {
+        status: 'current',
+        testFiles: []
+      })
+
+      expect([...plan.modules].sort()).toEqual([
+        'artifact_provenance',
+        'reviewer_orchestrator',
+        'session_persistence',
+        'workspace_page',
+        'workspace_runtime'
+      ])
+      expect(plan.reasonChains).toEqual(
+        expect.arrayContaining([
+          'reviewer_orchestrator -> workspace_runtime',
+          'reviewer_orchestrator -> workspace_page',
+          'reviewer_orchestrator -> artifact_provenance',
+          'reviewer_orchestrator -> artifact_provenance -> session_persistence'
+        ])
+      )
+      expect(plan.testFiles).toEqual(
+        expect.arrayContaining([
+          'src/main/artifacts/provenance-repository.architecture.test.ts',
+          'src/main/session-persistence/coordinator.architecture.test.ts',
+          'src/renderer/src/lib/acp/useWorkspaceAgentRuntime.architecture.test.ts',
+          'src/renderer/src/pages/workspace/workspace-page.architecture.test.ts'
+        ])
+      )
+    }
+  )
+
   it('fails closed for unknown and destructive changes', () => {
     expect(
       createAffectedTestPlan([{ path: 'src/main/unknown-owner.ts', status: 'added' }], {
